feat: add `.pipe(...)` for piping stdout of a command to another command
